Solved ships moving way too fast. Also removed a few unneeded cast to floats. Ships were moving too fast because the 'Mass' variable is inherited from 'GameplayObject'. I just had to have the ShipModule_Deserialize inherit 'GameplayObject' as well, and copy the 'Mass' read in from the XML files to it.
--HG--
branch : feature/Memory Usage Overhall

+++ Ship_Game/Gameplay/Ship.cs
@@ -3698,8 +3698,8 @@ namespace Ship_Game.Gameplay
                                 else
                                     this.ModulePowerDraw += moduleSlot.module.PowerDraw;
                                 this.Thrust += moduleSlot.module.thrust;
-                                this.WarpThrust += (float)moduleSlot.module.WarpThrust;
-                                this.TurnThrust += (float)moduleSlot.module.TurnThrust;
+                                this.WarpThrust += moduleSlot.module.WarpThrust;
+                                this.TurnThrust += moduleSlot.module.TurnThrust;
                                 if (moduleSlot.module.ECM > this.ECMValue)
                                 {
                                     this.ECMValue = moduleSlot.module.ECM;
